added and fixed some stuff in admin

+++ internal/events/user.go
@@ -17,10 +17,10 @@ type User struct {
 }
 
 type Member struct { //may use for nicks later
-	GuildId  int  `json:"guildId"`
-	Admin    bool `json:"admin"`
-	Owner    bool `json:"owner"`
-	UserInfo User `json:"userInfo"`
+	GuildId  int64 `json:"guildId"`
+	Admin    bool  `json:"admin"`
+	Owner    bool  `json:"owner"`
+	UserInfo User  `json:"userInfo"`
 }
 
 /*
